+++ datanode/sqlstore/migrations/0004_oracle_specs.sql
@@ -1,0 +1,4 @@
+-- +goose Up
+ALTER TABLE oracle_specs DROP COLUMN IF EXISTS signers;
+ALTER TABLE oracle_specs DROP COLUMN IF EXISTS filters;
+ALTER TABLE oracle_specs ADD COLUMN data JSONB NOT NULL;